refactor: project structure
update to match recommendations of Cosmic Python Chapter 4
