+++ src/allocation/settings/config.py
@@ -9,7 +9,7 @@ class LogConfig(BaseModel):
     """Logging configuration to be set for the server"""
 
     LOGGER_NAME: str = "allocation_service"
-    LOG_FORMAT: str = "%(levelprefix)s | %(asctime)s | %(message)s"
+    LOG_FORMAT: str = "[%(asctime)s] %(levelprefix)s %(message)s"
     LOG_LEVEL: str = "DEBUG"
 
     # Logging config
